Support CLIENT KILL from Redis 2.8.12+ in tests
CLIENT KILL since version 2.8.12 has a new format:
http://redis.io/commands/client-kill

The interesting part is `CLIENT KILL SKIPME yes/no`.
By default this option is set to yes, which means that something in the sense of client.kill(self) wouldn't work anymore.

This commit changes the test to use a different client for killing the desired connection, if the redis version is greater than 2.8.12